Fixed auth in info\acp_add_user.php
The auth in info\acp_add_user.php was wrong. This did not matter for
normal installation, but when I added a new link under Quick access on
the home screen of ACP every admin could access the module, even without
the specific permission. This fixes this, ensuring that the new link
only shows for users with the right permission. N.B. I did not (yet)
check if it is possible to bypass security by copy pasting urls